feat: add entity relations

+++ src/main/java/io/pakland/mdas/githubstats/domain/ports/CommitRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface CommitRepository extends JpaRepository<Commit,Integer> {
+public interface CommitRepository extends JpaRepository<Commit,Long> {
 }
